use the chunk to iterate records in the error case
pass the chunk to the error handler.  It will process each record
in the chunk in the same way that the plugin handles it, even
calling process_message() on a deep copy of the record.
All of the error handling has been moved into the error condition
in send_bulk, so that there is no error code in the regular,
non-error code path.  The only thing left is bulk_message_count
which must be done by the plugin and passed to the error
handling code.
ElasticsearchErrorHandler @records and @bulk_message_count are
now unused so I got rid of them.
I made sure that the tests covered all of the odd corner cases.